Basic Stack-Screen Navigation wired up. Home and Display Screens created, code moved out of App.js
